+++ R/ss_buc_factorial_anova_general.R
@@ -150,12 +150,13 @@
 ss_buc_factorial_anova_general <- function(F_observed, N, cells, df_numerator,
                                            df_denominator, alpha_prior = .05,
                                            alpha_planned = .05, assurance = .80,
-                                           power = .80, step = .001) {
-  v <- .validate_planning_inputs(alpha_prior, alpha_planned, assurance, power, step)
+                                           power = .80) {
+  v <- .validate_planning_inputs(alpha_prior, alpha_planned, assurance, power)
   alpha_prior <- v$alpha_prior
   alpha_prior_input <- v$alpha_prior_input
   assurance <- v$assurance
   power <- v$power
+  step <- v$step
 
   if (missing(N)) stop("You must specify 'N', which is the total sample size.")
   if (missing(cells)) stop("You must specify 'cells', the number of cells (groups) in the between-subjects design.")