+++ R/ss_buc_independent_t.R
@@ -144,12 +144,13 @@
 #'
 #' @template references
 ss_buc_independent_t <- function(t_observed, n, N, alpha_prior = .05, alpha_planned = .05,
-                           assurance = .80, power = .80, step = .001) {
-  v <- .validate_planning_inputs(alpha_prior, alpha_planned, assurance, power, step)
+                           assurance = .80, power = .80) {
+  v <- .validate_planning_inputs(alpha_prior, alpha_planned, assurance, power)
   alpha_prior <- v$alpha_prior
   alpha_prior_input <- v$alpha_prior_input
   assurance <- v$assurance
   power <- v$power
+  step <- v$step
 
   if (!missing(N)) {
     if (!is.null(N)) {