+++ R/ss_buc_paired_t.R
@@ -119,12 +119,13 @@
 #'
 #' @template references
 ss_buc_paired_t <- function(t_observed, N, alpha_prior = .05, alpha_planned = .05,
-                            assurance = .80, power = .80, step = .001) {
-  v <- .validate_planning_inputs(alpha_prior, alpha_planned, assurance, power, step)
+                            assurance = .80, power = .80) {
+  v <- .validate_planning_inputs(alpha_prior, alpha_planned, assurance, power)
   alpha_prior <- v$alpha_prior
   alpha_prior_input <- v$alpha_prior_input
   assurance <- v$assurance
   power <- v$power
+  step <- v$step
 
   if (missing(N)) stop("You need to specify a sample size (i.e., the number of pairs) used in the original study.")
   if (N <= 1) stop("Your total sample size is too small.")