+++ R/ss_buc_rm_anova_general.R
@@ -110,12 +110,13 @@
 #' @template references
 ss_buc_rm_anova_general <- function(F_observed, N, df_numerator, alpha_prior = .05,
                                 alpha_planned = .05, assurance = .80,
-                                power = .80, step = .001) {
-  v <- .validate_planning_inputs(alpha_prior, alpha_planned, assurance, power, step)
+                                power = .80) {
+  v <- .validate_planning_inputs(alpha_prior, alpha_planned, assurance, power)
   alpha_prior <- v$alpha_prior
   alpha_prior_input <- v$alpha_prior_input
   assurance <- v$assurance
   power <- v$power
+  step <- v$step
 
   if (missing(N)) stop("You must specify 'N', which is the total sample size.")
   if (missing(df_numerator)) stop("You must specify 'df_numerator', the numerator degrees of freedom for the effect of interest.")